Removed some unneeded code from the Client module Updated start script to cd to app directory before running start_fcgi script

+++ lib/DynIP/Controller/Client.pm
@@ -18,13 +18,10 @@ sub auto :Private {
 sub index :Path('/client') :Args(0) {
     my ( $self, $c ) = @_;
 
-    my $addr = $c->req->address;
-    my $server = $c->config->{name_server};
-    my $domain = $c->config->{domain};
-    my $fqdn = $c->model->fqdn($c->stash->{client});
-
-    $c->detach('/misconfigured')
-        unless defined $addr && defined $server && defined $domain && defined $fqdn;
+    unless (defined $c->req->address) {
+        $c->log->error("Could not get request address, something is terribly wrong!");
+        $c->detach('/misconfigured');
+    }
 
     $c->detach('/internal_error') 
         unless $c->model->update($c->stash->{client}, $c->req->address);
